Improve key tooltip in KeySelectionDialog
This adds a new field called "Status" into the tooltip which shows
whether the key is valid or not, and a reason why it's not valid. This
fixes the annoying problem when you select a key and it gets the  "bad key"
icon without any explanation as to why it is considered bad.

Secondly the tooltip is now placed in a table, which makes it look
a little prettier.