Add support for EncryptJobs with generic flags
This adds a new overloads for start and exec that instead of
distinct flags like always trust just accept any
GpgME::Context::EncryptFlags.

Backport from git.gnupg.org/gpgme 34b456c3f. git.gnupg.org/gpgme
also contains unit tests for this.

This is technically an ABI break but should not create problems
in practice as the new symbols are added at the bottom of the
vtable and there are no known implementers of the jobs outsite
of libkleo.