Introduce Kleo::DefaultKeyGenerationJob
The DefaultKeyGenerationJob takes only email and name, assembles the
XML with additional arguments using default key type (RSA), key length
(2048 bits) and passes it to the proper KeyGenerationJob.

The purpose of this job is to avoid having to have to build the nasty
XML everytime we want to generate a simple key pair.

Would be nice to eventually have a full-featured job that would
allow customizing more arguments, like key type, length, purpose
etc.