Simplify and split DirectoryServicesWidget
The old UI was probably written under the assumption that there
will be support for multiple OpenPGP Keyserver entries at some point.
We now only support a single entry for keyserver.
As GnuPG only supports a single Keyserver URL the old widget
lead to confusion and buggy behavior as it allowed the user to
configure multiple PGP Keyservers which were not respected or
could not be configured through gpgconf.

GnuPG since 2.1.16 has a built in keyserver URL, the placeholder
text in the line entry when no other keyserver is configured is
intended to reflect that.

API compatibility is kept with the old widget operation. Even though
I doubt that the read only settings are actually used in practice.

While this fixes bugs let's keep it in master because of regression
risks.

BUG: 192059
BUG: 359151