Fix crash in Keyselectioncombo
When switching the filterRegExp and quickly thereafter
opening the combobox there could be a crash in
QSortFilterProxyModel::mapToSource.
From my understanding this was because index.internalPointer
was invalid. As the documentation says that it should
be avoided to use the internal pointer I'm doing that now.

This fixes the crash and I did not see any regressing
behavior.

Only in master because the filterRegExp switching is
only something that the NewKeyApprovalDialog utilizes.

GnuPG-Bug-Id: T3861