Fix reenable of keyselectioncombo after keylisting
We only want to renenable the keyselectioncombo when we
triggered the keylisting in refresh keys and so disabled
the combo. Otherwise the enable / disable state is changed
at random times when any keylisting in the global keycache
finishes.

This fixes a bug in Kleopatra's signencryptwidget where
random keylistings would enable the combo box even if
the operation belonging to that combo was unchecked.