Fix key mapping based on summary column
If we move the summary column below num columns data queries
for this column fail. This breaks for example the
certificatelineedit in Kleopatra.

The summary column is still hidden because we do not set
header data.

GnuPG-Bug-ID: T4569

This fixes: 4847fcc27afc8101752de82b0dd1f5fee027695d

Merge is not neccessary as 4847fcc27 was not included
in any branch so far.