Change card API for gpgcard and use it
The idea is to leave the old API in Kleopatra around and
when GnuPG 2.3 is detected to use the new gpg-card API
in Libkleo.

The API is currently unstable and will only be stable
once merged into master. So the GpgCardWidget is thought
as a replacement of the SmartCardWidget for GnuPG 2.3
so that Kleopatra can switch between them.

The move into libkleo is done so that we can additionaly
write a standalone gpgcard GUI that will require GnuPG 2.3.