Take compliance mode name from libkleopatrarc
GnuPG is starting to get "compliance mode" users outside
of the German Government. As the approval for restricted
communication is also for NATO RESTRICTED and EU RESTRICTED
different users need to see different strings for the
mode.

The solution is now to use the names of the key filters already
defined in libkleopatrarc so that the string can be
changed through config.