Rename exception.h to kleoexception.h
On case-insensitive systems (Windows), the generated fancy
include Exception gets confused for the standard <exception>
include file, which then breaks builds in most curious
ways.

To avoid that, we rename exception.h to kleoexception.h, so the
fancy header is KleoException and thus the ambiguity is resolved.

As far as I checked the header is only included from Kleopatra,
which I will fix too.