Take care of immutable entries when updating a group of keys
Treat a group as immutable if the group itself or any of its entries is
immutable.

Reread the group after writing it to the configuration file to ensure
that it reflects the saved group in case of immutable entries. This
shouldn't really be necessary because immutable groups shouldn't be
editable by the user, but we do not enforce this ourselves.

GnuPG-bug-id: 5175, 5239