In mixed mode create "best of" list of resolved OpenPGP and S/MIME keys
If resolution with mixed protocols is allowed, then create a "best of"
list of the resolved OpenPGP and S/MIME keys, i.e. for a given recipient
take either the OpenPGP keys or the S/MIME keys where the choice if
based first on validity and then on the preferred protocol (with OpenPGP
as last resort).

GnuPG-bug-id: 5283