Replace all usages of the deprecated qBinaryFind
Where possible qBinaryFind is replaced with std::binary_search. All
other occurrences are replaced with a custom replacement.

I think using this replacement is much more convenient and results
in better readable and less errorprone code than replacing all usages of
qBinaryFind with std::lower_bound() and the necessary additional check
of the resulting iterator, in particular if a lambda expression is used
as custom comparison function.