Emit currentKeyChanged() signal even if it is already set
If the key that is set with setCurrentKey() became the current key by
changes in the underlying model (i.e. because it happens to be the item
at the current index), then no signal was emitted. This triggered
an update problem in the key approval dialog noticed after the previous
commit.

This problem is fixed by this change because on all model changes
updateWithDefaultKey() is called which in turn calls setCurrentKey().

GnuPG-bug-id: 5283