Always include the default key in the list of keys to chose from
Add custom subclass of KeyListSortFilterProxyModel that allows specifying
a key that is always accepted by the filtering. Use this subclass
to make sure that the default key is not filtered out.

This fixes the problem that KMail didn't show keys not matching
the email address of an identity in the identity configuration dialog
and that KMail cleared the keys silently if one saved such an identity
after opening the Cryptography tab.

Funded by: Intevation GmbH

BUG: 437091
FIXED-IN: 5.18.0