Remove special widget for OpenPGP keyserver entry
It is only used for the gpg/keyserver entry that is no longer made
available via gpgconf in GnuPG 2.3. Moreover, it tries to parse
the keyserver entry which fails for LDAP urls. And it silently keeps
the optional keyserver configuration options one can append to gpg's
(but not dirmngr's) keyserver option. But those option may no longer be
appropriate if the user changes the keyserver.

Instead of this special widget now a simple line edit is used which
isn't very user friendly, but, OTOH, it doesn't hide anything from the
user.

GnuPG-bug-id: 5404