Add workaround for broken tab order of dialog buttons
At least with the KdeLayout the Close button is excluded from the
forward tab order (but not from the backward tab order), so that the
Close button cannot be selected by pressing Tab repeatedly.

The buttons are displayed in the order "Copy to Clipboard", "Save to
Disk...", "Close", but nextInFocusChain() of the Save button returns the
AuditLogViewer instead of the Close button. The complete picture is
as follows:
- focusProxy() of the button box is the Copy button (okay)
- nextInFocusChain() of the Copy button is the Save button (okay)
- nextInFocusChain() of the Save button is the AuditLogViewer (not okay)
- nextInFocusChain() of the Close button is the Copy button (not okay)
Seems to be a bug in Qt.

Creating the Close button after the custom buttons works around the
problem and fixes the tab order (at least for KdeLayout).

GnuPG-bug-id: 5535