Add helpers for reading/writing key groups from/to a file
These helpers are mainly meant to be used for import and export of
key groups. Unlike KeyGroupConfig which uses KConfig, these helpers
use QSettings for reading/writing the groups from/to a file. QSettings
is used for reading because KConfig always reads global settings
additionally to the specified file. This could lead to unexpected
results, if any of the global settings files contains config groups
that are or look like key groups. For symmetry and to avoid format
problems, QSettings is also used for writing key groups to a file.

GnuPG-bug-id: 5638