Prevent cycles in graph of certificate issuers
Check the graph of certificate issuers for cycles when adding keys
to the hierarchical key list model. If a key would cause a cycle, then
mask its issuer causing it to be treated as top-level key. It is
undefined which key in a circular certificate chain will become the
top-level key of this chain.

Kudos to the Boost Graph Library!

GnuPG-bug-id: 5697