Check version of QGpgme instead of gpgme++
CryptoConfig is defined in QGpgme.

This should fix compilation if a newer version of gpgme++ is used with
an older version of QGpgme.
Note that, even if compilation against a mix of different versions of
gpgme, gpgme++, and QGpgme succeeds, this isn't supported and it's likely
that you will run into weird problems. Those three libraries are
released as a single tarball for a reason.