Allow new refresh job if the previous job is done
Emitting keyListingDone() may trigger changes to keys which may
require a new refresh before the just finished refresh job is
destroyed (its destruction is postponed with deleteLater()).
Therefore we clear the guarded pointer explicitly instead
of waiting for the job to be destroyed, so that a new refresh is started
even if the old refresh job is not yet destroyed.

GnuPG-bug-id: 5715