Explicitly exclude deprecated gpgsm/keyserver option from config UI
The "keyserver" option of gpgsm has been deprecated in favor of the
"ldapserver" option of dirmngr. Exclude the deprecated option from
the generic GnuPG configuration UI to avoid confusion and
misconfiguration.

This change also makes the lookup of entries to exclude independent
of the logical group of the config entries.

GnuPG-bug-id: 5801