Add helpers for more accessible dates and hex id strings
These helpers create strings that are more suitable for screenreaders
than the corresponing strings created for displaying on screen.

For dates we use the system's long format. Unfortunately, this may contain
the day of the week which is a bit too verbose, but it's certainly better
understandable than a string of three numbers.

For hex ids (i.e. key ids and fingerprints) we take groups of four
separating all characters with spaces (to avoid strings of digits to
be read as numbers) and separating all groups with commas (to force
a pause between the groups).

GnuPG-bug-id: 5841