Make the model used for the key selection combo a single-column model
If we don't do this, then QAccessibleTable, which is used for the
QListView of the combo box pop up, miscalculates the child from the
logical index. The causes the weird effect that, when you move the
cursor down in the pop up, then the screen reader reads the model
items in columns-first order, i.e. for the second entry in the
combo box it reads the second column of the first row (i.e. the first
entry) instead of the first column of the second row.
This is clearly a bug in QAccessibleTable which shouldn't consider the
column count if QAccessibleTable provides the interface for a list view.

For the combo box we anyway need the first column of the key list model
only, so we can simply pretend that there is only one column.

GnuPG-bug-id: 5845