Fix invalid regular expression
Simplify the regular expression that is used to split several match
contexts. The old regular expression contained "_-!" in a character set
which is invalid because '_' comes after '!'. QRegExp has accepted this
(but it's unclear what it made of the regular expression), but
QRegularExpression rejects it.

This fixes the warning
QString::split: invalid QRegularExpression object