Prevent screen readers from reading the tool tip
The tool tip shows details about the key using an HTML table. It is not
suitable for being read by screen readers. (orca even reads all the HTML
tags.) Since Qt uses the tool tip as fallback for the accessible
description if the accessible description is empty, we have to set a
non-empty accessible description to prevent Qt from doing so.

Alternatively, we could set an accessible description giving details
about the key suitable for screen readers, but the value of the combo
box already gives more than enough details.

GnuPG-bug-id: 6046