Add SystemInfo::isDarkModeActive
At least on Windows HighContrastMode does not automatically
mean dark mode. The problem is that the Dark Mode detection
in Qt does not seem to work properly so that we only handle
Dark Modes in High Contrast mode and not if the User enabled
a Dark Theme through AppsUseLightTheme registry value.

GnuPG-Bug-Id: T6076