Add ExpiryChecker
This is mostly a copy of NearExpiryChecker from pim/messagelib. The
code of the class is functionally identical. The handling of the
temporary GNUPGHOME for the unit test has been changed to the pattern
used for keyresolvercoretest.

I have readded the copyright information from keyresolver.* in
pim/messagelib from where the code has been factored out.

CCMAIL: sknauss@kde.org

GnuPG-bug-id: 6330