Add flag to request check of certificate chain
This makes it possible to request the check of the passed certificate
only. A check of the certificate chain must be requested explicitly.
Renamed KeyFlag(s) to CheckFlag(s) and OtherKey (= 0) to NoCheckFlags.
Also renamed circularCertificateChain test function to certificateChain
because it tests checking of certificate chains. Testing that circular
chains are detected is just a welcome side effect of the test data.

GnuPG-bug-id: 6330