Make expiration duration the actual days until/since expiry
People calculate in days and not in multiples of 24 hours, e.g. if
a key expired 1 day ago most humans will understand this as "the key
expired yesterday" and not as "the key expired 24-47:59 hours ago".

Also use "yesterday", "today", and "tomorrow" instead of "one day ago",
"in less than a day", and "in less than a day".

GnuPG-bug-id: 6330