Fix checking never expiring subkeys for expiration
A subkey that doesn't have an expiration date expires together with the
primary key (if the latter expires). This is how a default-generated
OpenPGP key looks like (i.e. expiring primary key, non-expiring subkey).

Additionally:
Add a sign-only OpenPGP test key to test the check for a suitable subkey
for OpenPGP.

GnuPG-bug-id: 6330