Deprecate DefaultKeyGenerationJob
GnuPG generates Ed25519 keys by default and has switched to 3072 bits as
default for RSA keys. DefaultKeyGenerationJob always generates RSA 2048
keys and uses the old parameter file API of GnuPG. It shouldn't be used
anymore.
It seems to be used by KMail in the Account Wizard and the identity
configuration.

GnuPG-bug-id: 6805