Deprecate Gpg4win class and add gpgvVerify instead
This adds a general function to use gpgv to verify a file
against a specific keyring. This deprecates the clunky
Gpg4win class / methods used previously for this and is more
general. And might even be useful in a lower tier.

GnuPG-Bug-Id: T6344
(cherry picked from commit 3822ae2ea499328272e0365a99cebcbcd8cb6eb3)