Add ScrollArea from Kleopatra
This adds the ScrollArea from Kleopatra under the name
AdjustingScrollArea. The ScrollArea has the advantage that with the
right sice policy it sets its size hint so that the scrollbars
can be avoided.
Users can then decide if they want to allow that or reduce the
size of the widgets so that scrollbars are visible.

GnuPG-Bug-Id: T6837