Add option to force a refresh of the key cache
While a refresh is running requests to refresh the cache are ignored and
listening for file system changes is disabled. This may cause externally
imported certificates to be missed. The option to force a cache refresh
allows users running a command that potentially adds new certificates
(e.g. "gpgsm --learn-card") to trigger a refresh to ensure all certificates
are loaded in the cache.

GnuPG-bug-id: 6846