Try to prevent some invalid LDAP servers
This blocks the user from adding LDAP servers that will result in empty hostnames;
This includes anything containing a port number or scheme in the hostname.
It does not ultimately prevent the user from finding ways of adding other ways
of adding invalid servers, or valid hostnames that don't have an LDAP server behind them.

GnuPG-Bug-ID: 6950