Fix showing keys as disabled / revoked / invalid
The existing code first checks whether the key is certified and then
whether other conditions apply, which means that the condition is never
shown for certified keys. For expired and revoked, this seems to still work
since the key is not considered certified if it's expired or revoked, but for
disabled keys, that's not the case.

GnuPG-Bug-ID: 7089