Fix columns not showing up in some cases
This can be reproduced e.g. in the "Imported certificates" tab: enabling a non-default column does not make it show up.
It's caused by the respective column's width being 0, which is restored from state cache. This used to work correctly as
the column widths were reset when a column was enabled, which was removed in 6745c9d3794e2de25cd8c252f2e3e3aafa5eb8a9

GnuPG-Bug-ID: 7210