Adjust column to its contents when it's made visible
When a hidden column is shown then Qt uses either the previously used
size (e.g. the size we set to the default size when we restore a hidden
column) or the default size. If a column has the default size after
being shown then it's very likely that the column wasn't resized to its
contents.

GnuPG-bug-id: 7227