Don't sort tree views/widgets unless sorting is enabled
If sorting isn't enabled for a tree view/widget then it must not be
sorted when restoring the column layout for the first time.

This unwanted initial sorting messed up the order of the card slots in
the new card keys view when a card of a some type was displayed for the
first time.

GnuPG-bug-id: 7018