Move ClipboardRole out of KeyList namespace
This role is useful for all kinds of models. The value is the highest
possible value with the same number of bits as Qt::UserRole.

I keep it in this file because I didn't find a more suitable file.

GnuPG-bug-id: 7018
(cherry picked from commit 6c15949fda8f4f612bef3e71c7f8baca594045fc)