Show key ID instead of short key ID in KeyRequester
In particular, this change gets rid of the wrong assumption that the
short key ID is always the last 4 octets of the fingerprint. This is
only true for V4 keys. By using Key::keyID we ensure that the correct
key ID is displayed also for V5 keys (and ancient V3 keys).

KeyRequester is used by KMail (and other apps using messagecomposer)
for displaying the keys used for encryption.

GnuPG-bug-id: 7297, 7301