Start gpgconf detached if we don't have an event loop
Waiting until gpgconf finishes turned out to be a bad idea because the
first start after a reboot of my Windows VM took more than 5 seconds
in 3 of 3 tries (resulting in a timeout on each try). Starting gpgconf
detached won't magically fix Windows, but at least we don't block the
startup of Kleopatra for longer than it takes.

GnuPG-bug-id: 6567