Fix look-up of libassuan and libgpg-error
We have to specify the paths found by pkg-config as HINTS for find_path
and find_library (instead of as PATHS) so that cmake considers those paths
before the default paths. Otherwise, we might get the header/library
provided by the distribution instead of those a .pc file in a custom
PKG_CONFIG_PATH points to.