Add helper to check if a high-contrast color scheme is in use
If a high-contrast color scheme is in use then we want to avoid the
usage of custom colors. Since we allow overriding the system's
high-contrast mode setting by configuring a custom color scheme we need
to check if the default scheme is in use in addition to checking the
status of the high-contrast mode.

GnuPG-bug-id: 7610