Avoid incorrect height of collapsed expander
If the content area is large then the height of the size hint of the
expanded widget can be smaller than the height of the size hint of the
collapsed widget plus the (maximum) height of the content area. In one
case this caused the header to be squeezed into a collapsed height of
3 pixels. This is fixed by remembering the height of the collapsed
widget.

GnuPG-bug-id: 6117