Notify users of KeyCache when keys are removed
Surprisingly, KeyCache::remove() didn't emit the keysMayHaveChanged
signal so that users of KeyCache were unaware that keys have been
removed from the cache (unless an update of the cache was triggered,
e.g. by the file system watcher). This unexpected behavior is fixed by
emitting the signal by default when keys are removed from the cache.
Optionally, the signal emission can be supressed, e.g. if the removal
happens as part of a larger change.

GnuPG-bug-id: 7863